Add indexing of users that requires user is signed in
Add tests, views and controller method to allow a signed in user to
view an index of users (all at this time!).  The menu to see users
is only displayed if the user is signed in.  If a direct URL is
provided and the user isn't logged in they are redirected to the sign
om page. The friendly forwarding then redirects to the protected page
that was originally requested.  Nice!  Next up is making sample users
(9.3.2).